search counter sale

+++ src/main/java/com/datn/meou/repository/AccountRepository.java
@@ -15,6 +15,6 @@ public interface AccountRepository extends JpaRepository<Account, Long> {
 
     Optional<Account> findByUsernameAndStatus(String username, Boolean status);
 
-    Optional<Account> findByIdAndStatus(Integer id, Boolean status);
+    Optional<Account> findByIdAndStatus(Long id, Boolean status);
 
 }
